Separated PublishListenListener
